refactor: move optionallySerializeAndSend out from server
