+++ packages/http-server/src/serializers.ts
@@ -1,9 +1,17 @@
+import { Dictionary } from '@stoplight/types';
 import { j2xParser } from 'fast-xml-parser';
+import { FastifyReply } from 'fastify';
+import { ServerResponse } from 'http';
 import typeIs = require('type-is');
+
+type Serializer = {
+  test: (contentType: string) => boolean;
+  serialize: (value: string | object) => string;
+};
 
 const xmlSerializer = new j2xParser({});
 
-export default [
+export const serializers = [
   {
     test: (value: string) => !!typeIs.is(value, ['application/json', 'application/*+json']),
     serialize: (data: any) => {
@@ -37,3 +45,25 @@ export default [
     },
   },
 ];
+
+export const optionallySerializeAndSend = (
+  reply: FastifyReply<ServerResponse>,
+  output: { headers?: Dictionary<string, string>; body?: string | object },
+  respSerializers: Serializer[],
+) => {
+  const contentType = output.headers && output.headers['Content-type'];
+  const serializer = respSerializers.find((s: Serializer) => s.test(contentType || ''));
+
+  const data = serializer && output.body ? serializer.serialize(output.body) : output.body;
+  const isInUTF8 = contentType && contentType.includes('charset=utf-8');
+
+  if (serializer) {
+    reply.serializer(serializer.serialize);
+  }
+
+  if (contentType && contentType.includes('application/json')) {
+    reply.header('content-type', isInUTF8 ? contentType : contentType + '; charset=utf-8');
+  }
+
+  reply.send(data);
+};
